SERVER-24363 cast mallinfo members to long long before adding

+++ src/mongo/util/processinfo_linux.cpp
@@ -451,7 +451,8 @@ void ProcessInfo::getExtraInfo(BSONObjBuilder& info) {
     struct mallinfo malloc_info =
         mallinfo();  // structure has same name as function that returns it. (see malloc.h)
     info.append("heap_usage_bytes",
-                malloc_info.uordblks /*main arena*/ + malloc_info.hblkhd /*mmap blocks*/);
+                static_cast<long long>(malloc_info.uordblks) /*main arena*/ +
+                    static_cast<long long>(malloc_info.hblkhd) /*mmap blocks*/);
     // docs claim hblkhd is included in uordblks but it isn't
 
     LinuxProc p(_pid);
